requirements.txt
fastapi
uvicorn
python-multipart
python-docx
openai
python-dotenv
jinja2
numpy
psycopg[binary]
pypdf
